// src/app/generated/vm-api/model/vmType.ts
/**
 * Player VM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type VmType = 'Unknown' | 'Vsphere' | 'Proxmox' | 'Azure';

export const VmType = {
    Unknown: 'Unknown' as VmType,
    Vsphere: 'Vsphere' as VmType,
    Proxmox: 'Proxmox' as VmType,
    Azure: 'Azure' as VmType
};